+++ src/connected/flux/atomicActions.ts
@@ -5,8 +5,9 @@ export enum ActionType {
   DATA_REQUESTED = 1,
   DATA_RETURNED,
   DATA_ERRORED,
-  SET_DATA_LOADER,
   SET_SERIES_IDS,
+  SET_DATA_LOADER,
+  SET_DATA_LOADER_DEBOUNCE_TIMEOUT,
   SET_X_DOMAIN,
   SET_OVERRIDE_X_DOMAIN,
   SET_Y_DOMAINS,
@@ -31,6 +32,7 @@ function createActionCreator<P>(type: ActionType) {
 
 export const setSeriesIds = createActionCreator<SeriesId[]>(ActionType.SET_SERIES_IDS);
 export const setDataLoader = createActionCreator<DataLoader>(ActionType.SET_DATA_LOADER);
+export const setDataLoaderDebounceTimeout = createActionCreator<number>(ActionType.SET_DATA_LOADER_DEBOUNCE_TIMEOUT);
 export const setChartPhysicalWidth = createActionCreator<number>(ActionType.SET_CHART_PHYSICAL_WIDTH);
 
 export const setXDomain = createActionCreator<Interval>(ActionType.SET_X_DOMAIN);