+++ src/connected/flux/reducer.ts
@@ -56,6 +56,11 @@ export default function(state: ChartState, action: Action<any>): ChartState {
         dataLoader: { $set: action.payload || invalidLoader }
       });
 
+    case ActionType.SET_DATA_LOADER_DEBOUNCE_TIMEOUT:
+      return update(state, {
+        debounceTimeout: { $set: _.isNumber(action.payload) ? action.payload : state.debounceTimeout }
+      });
+
     case ActionType.SET_CHART_PHYSICAL_WIDTH:
       return update(state, {
         physicalChartWidth: { $set: action.payload }